@@ -1,18 +1,17 @@
-const exec = require('@actions/exec')
+import * as exec from '@actions/exec'
 
-const execute = async command => {
+export const execute = async command => {
   let log = ''
-  const options = {}
-  options.listeners = {
-    stdout: data => {
-      log += data.toString()
-    },
-    stderr: data => {
-      log += data.toString()
+  const options = {
+    listeners: {
+      stdout: data => {
+        log += data.toString()
+      },
+      stderr: data => {
+        log += data.toString()
+      },
     },
   }
   await exec.exec(command, null, options)
   return log
 }
-
-module.exports = execute
